feat: enhance counters and sub-categories modules with user authentication, improve error handling, and refactor service methods

+++ src/counters/counters.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateCounterDto } from './dto/create-counter.dto';
 import { UpdateCounterDto } from './dto/update-counter.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -7,7 +7,15 @@ import { PrismaService } from 'src/prisma/prisma.service';
 export class CountersService {
   constructor(private readonly prisma: PrismaService) {}
 
-  create(createCounterDto: CreateCounterDto) {
+  async create(createCounterDto: CreateCounterDto) {
+    const existingCounter = await this.prisma.counter.findUnique({
+      where: { name: createCounterDto.name },
+    });
+
+    if (existingCounter) {
+      throw new BadRequestException('El contador ya existe');
+    }
+
     return this.prisma.counter.create({
       data: createCounterDto,
     });
@@ -27,14 +35,30 @@ export class CountersService {
     });
   }
 
-  update(id: number, updateCounterDto: UpdateCounterDto) {
+  async update(id: number, updateCounterDto: UpdateCounterDto) {
+    const existingCounter = await this.prisma.counter.findUnique({
+      where: { id },
+    });
+
+    if (!existingCounter) {
+      throw new BadRequestException('El contador no existe');
+    }
+
     return this.prisma.counter.update({
       where: { id },
       data: updateCounterDto,
     });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const counter = await this.prisma.counter.findUnique({
+      where: { id },
+    });
+
+    if (!counter) {
+      throw new BadRequestException('El contador no existe');
+    }
+
     return this.prisma.counter.delete({
       where: { id },
     });
